+++ lib/src/utils/mapper_utilities.dart
@@ -22,14 +22,18 @@ class MapUtils {
 
   /// Maps the value of the future to a new value.
   static Future<S> mapAsync<T, S>(
-      {required T data, required Mapper<T, S> mapper}) async {
+      {required T data,
+      required Mapper<T, S> mapper,
+      bool printError = kDebugMode}) async {
     try {
       return mapper(data);
     } catch (e, s) {
-      PlayxLogger.printError(
-          header: 'MapAsync Error',
-          text: e.toString(),
-          stackTrace: s.toString());
+      if (printError) {
+        PlayxLogger.printError(
+            header: 'MapAsync Error',
+            text: e.toString(),
+            stackTrace: s.toString());
+      }
       // Handle any errors in the main thread
       rethrow;
     }
@@ -38,20 +42,24 @@ class MapUtils {
   /// Maps the value of the future to a new value in an isolate.
   /// If [useWorkManager] is set to true, it will use [workerManager] to execute the task.
   /// Otherwise, it will use [compute] to execute the task.
-  static Future<S> mapAsyncInIsolate<T, S>(
-      {required T data,
-      required Mapper<T, S> mapper,
-      bool useWorkManager = true}) async {
+  static Future<S> mapAsyncInIsolate<T, S>({
+    required T data,
+    required Mapper<T, S> mapper,
+    bool useWorkManager = true,
+    bool printError = kDebugMode,
+  }) async {
     try {
       final res = useWorkManager
           ? await workerManager.execute(() => _mapAsync<T, S>([data, mapper]))
           : await compute(_mapAsync<T, S>, [data, mapper]);
       return res;
     } catch (e, s) {
-      PlayxLogger.printError(
-          header: 'MapAsyncInIsolate Error',
-          text: e.toString(),
-          stackTrace: s.toString());
+      if (printError) {
+        PlayxLogger.printError(
+            header: 'MapAsyncInIsolate Error',
+            text: e.toString(),
+            stackTrace: s.toString());
+      }
       // Handle errors occurring in the isolate
       rethrow;
     }
@@ -60,14 +68,23 @@ class MapUtils {
 
 extension MapAsync<T> on T {
   /// Maps the value of the T to a new value.
-  Future<S> mapAsync<S>({required Mapper<T, S> mapper}) =>
-      MapUtils.mapAsync(data: this, mapper: mapper);
+  Future<S> mapAsync<S>({
+    required Mapper<T, S> mapper,
+    bool printError = kDebugMode,
+  }) =>
+      MapUtils.mapAsync(data: this, mapper: mapper, printError: printError);
 
   //// Maps the value of the [T] to a new value in an isolate.
   /// If [useWorkManager] is set to true, it will use [WorkerManager] to execute the task.
   /// Otherwise, it will use [compute] to execute the task.
-  Future<S> mapAsyncInIsolate<S>(
-          {required Mapper<T, S> mapper, bool useWorkManager = true}) =>
+  Future<S> mapAsyncInIsolate<S>({
+    required Mapper<T, S> mapper,
+    bool useWorkManager = true,
+    bool printError = kDebugMode,
+  }) =>
       MapUtils.mapAsyncInIsolate(
-          data: this, mapper: mapper, useWorkManager: useWorkManager);
+          data: this,
+          mapper: mapper,
+          useWorkManager: useWorkManager,
+          printError: printError);
 }